Move CancelPurchase render functions into sub components
